# app/v1/swagger.yml
...
swagger: '2.0'
...
info:
  version: '1.0.0'
  title: N0deJS API
  description: N0deJS REST API prototype
schemes:
  - http
host: localhost:3001
basePath: /v1
...
paths:
  '/deployments/{deploymentId}':
    get:
      summary: Get a deployment with deployment id specified in the request.
      description: |
        Get a deployment with deployment id specified in the request.
        If there is no deployment, returns 404.
      operationId: getDeployment
      consumes:
        - application/json # This operation doesn't actually consume anything, this should be removed?
      produces:
        - application/json
      # This is array of GET operation parameters:
      parameters:
        - name: deploymentId
          in: path
          description: The deployment id to find and return deployment.
          required: true
          type: integer
          format: int32
      # Expected responses for this operation:
      responses:
        # Response code
        200:
          description: Successful response with a deployments.
          # A schema describing the response object.
          schema:
            $ref: '#/definitions/deployment'

    delete:
      summary: Delete the deployment specified in the request.
      description: |
        Delete the deployment specified in the request.
        If there is no deployment, returns 404.
      operationId: deleteDeployment
      consumes:
        - application/json # This operation doesn't actually consume anything, this should be removed?
      produces:
        - application/json
      # This is array of GET operation parameters:
      parameters:
        - name: deploymentId
          in: path
          description: The deployment id to find and return deployment.
          required: true
          type: integer
          format: int32
      # Expected responses for this operation:
      responses:
          # Response code
          404:
            description: Deployment with specified id is not found.
          # Response code
          204:
            description: Deployment Succefully deleted.
...
  '/projects/{projectId}/deployments':
    get:
      summary: Get the set of deployments associated with the project specified in the request.
      description: |
        Get the set of deployments associated with the project specified in the request.
        If there are no deployments, returns an empty list.
      operationId: getDeploymentsForProject
      consumes:
        - application/json # This operation doesn't actually consume anything, this should be removed?
      produces:
        - application/json
      # This is array of GET operation parameters:
      parameters:
        - name: projectId
          in: path
          description: The project id to list deployments for.
          required: true
          type: integer
          format: int32
      # Expected responses for this operation:
      responses:
        # Response code
        200:
          description: Successful response with a listing of deployments.
          # A schema describing the response object.
          schema:
            $ref: '#/definitions/deployments'

    post:
      summary: Create a deployment for a project.
      description: |
        Creates a new deployment for the specified project and return 201.
        If failed to create returns 404.
      operationId: createDeployment
      consumes:
        - application/json # This operation doesn't actually consume anything, this should be removed?
      produces:
        - application/json
      # This is array of GET operation parameters:
      parameters:
        - name: projectId
          in: path
          description: The project id to create post deployment actions for.
          required: true
          type: integer
          format: int32

        - name: deployment
          in: body
          description: The deployment JSON to post.
          schema:
            $ref: '#/definitions/deployment'
          required: true

      # Expected responses for this operation:
      responses:
        # Response code
        201:
          description: Successful response with details of the newly created post deployment.
          # A schema describing the response object.
          schema:
            $ref: '#/definitions/deployment'
...
  '/targets/{targetId}/':
    # This is a HTTP operation
    get:
      summary: Get the deployment environment/target specified by `targetId`.
      description: |
        Get the deployment environment/target specified in the request. If the
        item is not found, return a `404/NOT_FOUND`.
      operationId: getEnvironment
      consumes:
        - application/json # This operation doesn't actually consume anything, this should be removed.
      produces:
        - application/json
      # This is array of GET operation parameters:
      parameters:
        - name: targetId
          in: path
          description: The internal target id.
          required: true
          type: integer
          format: int32
      # Expected responses for this operation:
      responses:
        # Response code
        200:
          description: Successful response with a target/environment object
          # A schema describing your response object.
          schema:
            $ref: '#/definitions/target'
        # Response code
        404:
          description: The specified target/environment is not found.

    put:
      summary: Update the deployment environment specified by `targetId`.
      description: |
        Update the specified deployment environment and return the updated environment.
      operationId: updateEnvironment
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: targetId
          in: path
          description: The id of the environment.
          required: true
          type: integer
          format: int32

        - name: target
          required: true
          in: body
          description: The target object.
          schema:
            $ref: '#/definitions/target'

      # Expected responses for this operation:
      responses:
        # Response code
        200:
          description: The updated target.
          schema:
            $ref: '#/definitions/target'
        # Response code
        default:
          description: Failed to update a target.
          schema:
            $ref: '#/definitions/error'

    # This is a HTTP operation
    delete:
      summary: Delete the deployment environment/target specified by `targetId`.
      description: |
        Delete the deployment environment/target referenced in the request and return 204.
        If the target is not found, returns `404`.
      operationId: deleteEnvironment
      consumes:
        - application/json # This operation doesn't actually consume anything, this should be removed.
      produces:
        - application/json
      # This is array of GET operation parameters:
      parameters:
        - name: targetId
          in: path
          description: The environment id we want to delete.
          required: true
          type: integer
          format: int32
      # Expected responses for this operation:
      responses:
        # Response code
        500:
          description: When environment with specified id is not found or there are projects using it.
        # Response code
        204:
          description: Successfully deleted.
...
definitions:
...
  deployment:
    type: object
    description: Models a deployment.
    required:
      - id
      - name
      - createdDate
      - buildId
      - projectId
    properties:
      id:
        type: integer
        format: int32
        description: The internal deployment id.
      buildId:
        type: integer
        format: int32
        description: The build id to be deployed.
      projectId:
        type: integer
        format: int32
        description: The project id to be deployed.
      name:
        type: string
        description: The name of the deployment.
      createdDate:
        type: string
        description: Date time when deployment is created.
      browseUrl:
        type: string
        description: The url where the build is deployed.


  ############## Model: deployments ##############
  deployments:
    type: object
    description: |
      The Deployments for a project. This object wraps an array of `deployment` objects.
    required:
      - deployments
    properties:
      deployments:
        type: array
        description: An array of `Deployment` objects, which can be empty.
        items:
          $ref: '#/definitions/deployment'


  ############## Model: error ##############
  error:
    type: object
    description: |
      A standard error object returned by all API calls. TBD: this should be flattened into a single object,
      rather than having a nested `error` object.
    required:
      - apiVersion
      - error
    properties:
      apiVersion:
        type: string
        description: The API version, e.g. `v1`.
      error:
        type: object
        description: A standard error object returned by all API calls.
        required:
          - code
          - message
        properties:
          code:
            type: string
            description: The error code, usually an integer, often a HTTP status code.
          message:
            type: string
            description: The message associated with the error.
          log:
            type: string
            description: Any additional error information, such as log entries, stack trace, etc. Likely to be turned off in production environments.


  ############## Model: member ##############
...
  target:
    type: object
    description: Models a deployment target/environment (e.g. a CloudFoundry instance) that a project can be deployed to.
    required:
      - id
      - type
      - url
      - userName
      - password
      - name
      - organization
      - space
    properties:
      id:
        type: integer
        format: int32
        description: The internal id of this target instance.
      'type':
        type: string
        description: The type of target environment, e.g. `cloudfoundry`, `aws`, etc..
      url:
        type: string
        description: The URL endpoint that the target is accessible at.
      userName:
        type: string
        description: The username to authenticate to the target with.
      'password':
        type: string
        description: The password to authenticate to the target with.
      name:
        type: string
        description: The user-provided label for this target. # This field should be renamed to label
      organization:
        type: string
        description: The organization under which a project will be deployed to this target.
      space:
        type: string
        description: The space within an organization that a project will be deployed under on this target.


  ############## Model: targets ##############
...